requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
discord==1.0.1
idna==3.1
multidict==5.1.0
pyparsing==2.4.7
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
